Bugfix/circular import and sprk (#221)
* import fix

* spark version updated and uv implemented.
